Move keep and omit transforms to searchAndReplace. Update tests.
