+++ src/tests/converter.spec.ts
@@ -2,7 +2,6 @@ import { writeFile } from "fs/promises";
 import * as addConverter from "../converters/addTextTransform.js";
 import * as converter from "../converters/converter.js";
 import * as dateTransformFunctions from "../converters/dateTransform.js";
-import * as keepOrOmitTransforms from "../converters/keepOrOmit.js";
 import * as numericTransformFunctions from "../converters/numericTransform.js";
 import * as searchAndReplaceTransformFunctions from "../converters/searchAndReplace.js";
 import * as truncateTransformFunctions from "../converters/truncateTransform.js";
@@ -94,8 +93,8 @@ const spyOnGenerateRenameList = jest.spyOn(converter, "generateRenameList"),
     .mockImplementation(({ transforms, sourcePath }) =>
       Promise.resolve(mockRenameListToolSet.mockRollback)
     ),
-  spyOnKeepTransform = jest.spyOn(keepOrOmitTransforms, "keepTransform"),
-  spyOnOmitTransform = jest.spyOn(keepOrOmitTransforms, "omitTransform");
+  spyOnKeepTransform = jest.spyOn(searchAndReplaceTransformFunctions, "keepTransform"),
+  spyOnOmitTransform = jest.spyOn(searchAndReplaceTransformFunctions, "omitTransform");
 
 [
   spyOnDateTransform,
